Task 2/requirements.txt
attrs==19.3.0
big-O-dmunguiatec==0.0.1
cycler==0.10.0
importlib-metadata==1.0.0
kiwisolver==1.1.0
matplotlib==3.1.2
more-itertools==8.0.0
numpy==1.17.4
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.1
six==1.13.0
wcwidth==0.1.7
zipp==0.6.0
